@@ -15,15 +15,14 @@
         "bootstrap": "^5.1.3",
         "cross-env": "^7.0.3",
         "jquery": "^3.2",
-        "laravel-mix": "^6.0.13",
+        "laravel-mix": "^6.0.0-beta.17",
         "lodash": "^4.17.19",
         "popper.js": "^1.12",
-        "postcss": "^8.1.14",
+        "postcss": "^8.3.1",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.5",
-        "vue-template-compiler": "^2.6.12"
+        "vue": "^3.2.36",
+        "vue-loader": "^17.0.0"
     }
 }
