@@ -1,25 +1,17 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!-- CSRF Token -->
-        <meta name="csrf-token" content="{{ csrf_token() }}">
-        <title>{{ config('app.name', 'Vue Laravel SPA') }}</title>
-        <!-- Styles -->
-        <link href="{{ mix('/css/app.css') }}" rel="stylesheet">
+        <title>Laravel</title>
+        <script src="{{ asset('js/app.js') }}" defer="defer"></script>
     </head>
 
     <body>
         <div id="app">
-            <example-component></example-component>
+            <test-vue></test-vue>
         </div>
-
-        <!-- Scripts -->
-        <script src="{{ mix('/js/app.js') }}" defer></script>
-
     </body>
-
 
 </html>
